foi testado e insere nas pickingLists, foi muito dificil, se testarem chamar e n der digam porque eu adicionei a ligação à db primeiro atraves de tools->connect database

+++ Lib_Primavera/Model/LinhaDaPickList.cs
@@ -16,7 +16,7 @@ namespace FirstREST.Lib_Primavera.Model
 
 
 
-        public int Quantidade { get; set; }
+        public double Quantidade { get; set; }
 
         public string Localizacao { get; set; }
